% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model Checking.R
\name{AICBIC}
\alias{AICBIC}
\title{AIC and BIC of Specified HMM}
\usage{
AICBIC(HMM, observations)
}
\arguments{
\item{HMM}{An object of class \code{HMM} or \code{xHMM}, a hidden Markov model.}

\item{observations}{A numeric vector containing the observed data.}
}
\value{
A list containing \code{AIC} and \code{BIC}.
}
\description{
calculates the AIC and BIC of class HMM or xHMM.
}
