% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algorithm.R
\name{ForwardBackwardAlgorithm}
\alias{ForwardBackwardAlgorithm}
\title{Forward and Backward Algorithm for HMMs}
\usage{
ForwardBackwardAlgorithm(HMM, observations)
}
\arguments{
\item{HMM}{object of class \code{HMM} or \code{xHMM} containing the parameters
of the HMM.}

\item{observations}{a numeric vector of observations.}
}
\value{
Returns a list with elements:
\describe{
\item{Alpha}{a matrix containing the forward probabilities with time along the
columns and hidden states along the rows.}
\item{Alpha_scaled}{contains a scaled version of the forward probabilities.}
\item{Beta}{a matrix containing the backward probabilities with time along the
columns and hidden states along the rows.}
\item{Beta_scaled}{contains a scaled version of the backward probabilities.}
\item{Likelihood_forw}{a numeric containing the likelihood function evaluated
in \code{observation} and parameters from \code{HMM}. Computed using the forward
probabilities.}
\item{Likelihood_back}{a numeric containing the likelihood function evaluated
in \code{observation} and parameters from \code{HMM}. Computed using the backward
probabilities. }
\item{P_mat}{a matrix containing the state probabilities with time along the columns
and hidden states along the rows.}
}
}
\description{
Runs the forward and backward algorithm for an HMM and observation
to calculate the forward and backward probabilities.
}
\details{
The function calculates the forward probabilities,
\deqn{\alpha_t(i) = P(c_t = i, x_1,\dots,x_t),}
and the backward probabilities,
\deqn{\beta_t(i) = P(x_{t+1},\dots,x_T\vert c_t=i) \quad t=1,\dots,T-1,}
where \eqn{x_1,\dots,x_T} denotes the observed sequence and \eqn{c_1,\dots,c_T}
denotes the hidden Markov chain.

The forward or the backward probabilities can be used to calculate the likelihood
function and the state probabilities, i.e. \eqn{P(c_t=i\vert x)}.
}
