% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{HMM}
\alias{HMM}
\title{Hidden Markov Models}
\usage{
HMM(distribution, emis_param, initial_probs, TPM)
}
\arguments{
\item{distribution}{a character of length one specifying the emission distribution,
must be one of the following: "Poisson", "Bernoulli", "Normal".}

\item{emis_param}{the parameters of the emission distribution.
For "Poisson": must be a numeric vector or list of length one containing the rates.
For "Bernoulli": must be a numeric vector or list of length one containing the probabilities.
For "Normal": must be a list with two elements named "mean" and "sd".}

\item{initial_probs}{a probability vector with the initial distribution of the hidden Markov chain.}

\item{TPM}{a transition probability matrix for the hidden Markov chain.}
}
\value{
\code{HMM} returns an object of class \code{c("PoissonHMM", "HMM")},
\code{c("BernoulliHMM", "HMM")}, or \code{c("NormalHMM", "HMM")}.

An object of class \code{HMM} is a list containing the following elements:
\describe{
\item{\code{nstates}}{a numeric of length one. The size of the hidden state space.}
\item{\code{distribution}}{a character of length one. The emission distribution.}
\item{\code{emis_param}}{a list containing the parameters of the emission
distribution named in accordance with the parameter names used in base R's
functions for the probability distributions.}
\item{\code{initial_probs}}{a numeric vector of length \code{nstates}. The
initial probabilities of the hidden Markov chain.}
\item{\code{TPM}}{a square matrix with dimension \code{nstates}. The transition
probabilities of the hidden Markov chain.}
}
}
\description{
Creates an object of class \code{HMM} containing the parameters
of a hidden Markov model with a standard emission distribution.
}
\details{
An object of class \code{HMM} also has a subclass, which is either
\code{PoissonHMM}, \code{BernoulliHMM}, or \code{NormalHMM} depending on the
emission distribution specified in \code{distribution}.
}
\seealso{
\code{\link{is.HMM}}, \code{\link{is.PoissonHMM}},
\code{\link{is.BernoulliHMM}}, and \code{\link{is.NormalHMM}}
for class checking.
}
