% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algorithm.R
\name{HMM_EM}
\alias{HMM_EM}
\title{Expectation Maximization Algorithm for HMM Objects}
\usage{
HMM_EM(HMM, ...)

## S3 methods for PoissonHMM, BernoulliHMM, NormalHMM:
HMM_EM(HMM, observations, ...)
}
\arguments{
\item{HMM}{an object of class \code{HMM} containing model and initial parameters.}

\item{observations}{a numeric vector containing the observed sequence.}
}
\value{
Returns an object of class \code{HMM} containing the estimated parameters.
}
\description{
The EM-algorithm for objects of class \code{HMM}.
}
\details{
The EM-algorithm, also called the Baum-Welch algorithm in the HMM context,
can be used to find approximate maximum likelihood estimates for the parameters.
The desired model, i.e. number of hidden states and emission distribution, is
specified in \code{HMM}, which contains this specification along with initial
parameters for the EM-algorithm.
}
\seealso{
\code{\link{Nstate_HMM}} for easy fitting of an HMM with a standard emission
distribution.
}
