% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{is.BernoulliHMM}
\alias{is.BernoulliHMM}
\title{Class Checking for BernoulliHMM}
\usage{
is.BernoulliHMM(obj)
}
\description{
Checks whether an object inherits from class \code{BernoulliHMM}.
}
\seealso{
\code{\link{HMM}} for object construction, and \code{\link{is.PoissonHMM}},
\code{\link{is.NormalHMM}} for class checking.
}
