% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{is.HMM}
\alias{is.HMM}
\title{Class Checking for HMM}
\usage{
is.HMM(obj)
}
\description{
Checks whether an object inherits from class \code{HMM}.
}
\seealso{
\code{\link{HMM}} for object construction, and \code{\link{is.PoissonHMM}},
\code{\link{is.BernoulliHMM}}, \code{\link{is.NormalHMM}} for class checking.
}
