% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{is.NormalHMM}
\alias{is.NormalHMM}
\title{Class Checking for NormalHMM}
\usage{
is.NormalHMM(obj)
}
\description{
Checks whether an object inherits from class \code{NormalHMM}.
}
\seealso{
\code{\link{HMM}} for object construction, and \code{\link{is.PoissonHMM}},
\code{\link{is.BernoulliHMM}}, for class checking.
}
