% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{is.PoissonHMM}
\alias{is.PoissonHMM}
\title{Class Checking for PoissonHMM}
\usage{
is.PoissonHMM(obj)
}
\description{
Checks whether an object inherits from class \code{PoissonHMM}.
}
\seealso{
\code{\link{HMM}} for object construction, and
\code{\link{is.BernoulliHMM}}, \code{\link{is.NormalHMM}} for class checking.
}
