% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{is.xHMM}
\alias{is.xHMM}
\title{Class Checking for xHMM}
\usage{
is.xHMM(obj)
}
\description{
Checks whether an object inherits from class \code{xHMM}.
}
\seealso{
\code{\link{xHMM}} for object construction.
}
