% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_Decoding.R
\name{Local_decoding}
\alias{Local_decoding}
\title{Local Decoding}
\usage{
Local_decoding(HMM, observations)
}
\arguments{
\item{HMM}{Your specified hidden Markov model by using the function \code{HMM}
or \code{xHMM}}

\item{observations}{Your Data}
}
\value{
Vector of probability of being in each state
}
\description{
Calculate the probability of being in each state to time
time T.
}
