% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_Decoding.R
\name{Local_decoding_with_scaling}
\alias{Local_decoding_with_scaling}
\title{Local_decoding_with_scaling}
\usage{
Local_decoding_with_scaling(HMM, observations)
}
\arguments{
\item{HMM}{Your specified hidden Markov model by using the function \code{HMM}
or \code{xHMM}}

\item{observations}{Your Data}
}
\value{
Vector of probability of being in each state with scaling
}
\description{
Calculate the probability of being in each state to time
time T with scaling.
}
