% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nbinom2State.R
\docType{data}
\name{Nbinom2State}
\alias{Nbinom2State}
\title{Nbinom2State Dataset}
\format{
a numeric vector of length 100.
}
\usage{
Nbinom2State
}
\description{
An observation of length 100 simulated from a two-state negative binomial HMM.
}
\details{
The data is simulated with the parameters:
\describe{
\item{sizes:}{10, 100}
\item{probs:}{0.25, 0.30}
\item{initial probability for state one:}{0.5}
\item{transition probability from one to one:}{0.6}
\item{transition probability from two to two:}{0.8}
}
}
\keyword{datasets}
