% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_states_fit.R
\name{Nstate_HMM}
\alias{Nstate_HMM}
\title{Fit n-State HMM}
\usage{
Nstate_HMM(distribution = "Normal", data, nstates = 2)
}
\arguments{
\item{distribution}{a character of length one specifying the emission
distribution, must be one of the following: "Poisson", "Bernoulli", "Normal".}

\item{data}{a numeric vector containing the observed data.}

\item{nstates}{a numeric of length one, the size of the hidden state space.}
}
\value{
Returns an object of class \code{HMM} containing parameter estimates
for the fitted model.
}
\description{
Fit an HMM with a standard emission distribution and a specified
number of hidden states.
}
\details{
\code{Nstate_HMM} calls \code{HMM_EM} with initial parameters chosen
using summary statistics of \code{data}.
}
\seealso{
\code{\link{HMM_EM}} for the EM-algorithm.
}
