% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model Checking.R
\name{Nstates_AICBIC}
\alias{Nstates_AICBIC}
\title{Calculating AIC and BIC for Different State Space Sizes}
\usage{
Nstates_AICBIC(distribution = "Normal", data, nstates = 3, doplot = "ggplot")
}
\arguments{
\item{distribution}{a character of length one specifying the emission distribution,
must be one of the following: "Poisson", "Bernoulli", "Normal".}

\item{data}{a numeric vector containing the observed data.}

\item{nstates}{the number of states up to which the function should fit using the function \code{\link{nstates_HMM}}.
i.e. for \code{nstates=5} the data is fit and the AIC and BIC calculated for states 2 up to 5.}

\item{doplot}{can be specified as either "ggplot" or "base" according to which plot is wanted.
otherwise no plot will be made.}
}
\value{
Returns a \code{data.frame} containing AIC and BIC for HMM with the
specified \code{distribution} and stationary initial distribution for the
hidden Markov chain, when fitted to the data for 2 states up to \code{nstates}
states. Furthermore a plot is created in accordance with what has been specified
by \code{doplot}.
}
\description{
Calculating AIC and BIC for Different State Space Sizes
}
