% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson2State.R
\docType{data}
\name{Poisson2State}
\alias{Poisson2State}
\title{Poisson2State Dataset}
\format{
a numeric vector of length 100.
}
\usage{
Poisson2State
}
\description{
An observation of length 100 simulated from a two-state Poisson HMM.
}
\details{
The data is simulated with the parameters:
\describe{
\item{rates:}{10, 20}
\item{initial probability for state one:}{0.5}
\item{transition probability from one to one:}{0.6}
\item{transition probability from two to two:}{0.8}
}
}
\keyword{datasets}
