% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{xHMM}
\alias{xHMM}
\title{Hidden Markov Models}
\usage{
xHMM(distribution, emis_param, emis_paramspace, initial_probs, TPM)
}
\arguments{
\item{distribution}{a function, the pdf or pmf of the emission distribution.}

\item{emis_param}{the parameters of the emission distribution.}

\item{emis_paramspace}{a list containing the parameter spaces
of the emission parameters. Elements should be named according to the argument
name in \code{distribution} and contain precisely one of the characters:
"real", "pos_real", "prob"}

\item{initial_probs}{a probability vector with the initial distribution of the hidden Markov chain.}

\item{TPM}{a transition probability matrix for the hidden Markov chain.}
}
\value{
\code{xHMM} returns an object of class \code{xHMM}.

An object of class \code{xHMM} is a list containing the following elements:
\describe{
\item{\code{nstates}}{a numeric of length one. The size of the hidden state space.}
\item{\code{distribution}}{a function. The emission distribution.}
\item{\code{emis_param}}{a list containing the parameters of the emission
distribution named in accordance with the parameter names used in the arguments
of \code{distribution}.}
\item{\code{emis_paramspace}}{a list containing the parameter spaces
of the emission parameters. Elements are named according to the argument
name in \code{distribution}.}
\item{\code{initial_probs}}{a numeric vector of length \code{nstates}. The
initial probabilities of the hidden Markov chain.}
\item{\code{TPM}}{a square matrix with dimension \code{nstates}. The transition
probabilities of the hidden Markov chain.}
}
}
\description{
Creates an object of class \code{xHMM} containing the parameters
of a hidden Markov model (HMM) with a user specified emission distribution.
}
\seealso{
\code{\link{is.xHMM}} for class checking and \code{\link{HMM}} for
Poisson, Bernoulli, or Normal HMMs.
}
